test(check): split package registration memo test into its own module
